Check if "make" has been done.

+++ platformAWS/AmazonWebServices_ArangoDB_Cluster.sh
@@ -10,6 +10,12 @@
 #   NUMBER  : count of machines to create (e.g. -n 3)
 #   OUTPUT  : local output log folder (e.g. -d /my/directory)
 
+if [ platformAWS/AmazonWebServices_ArangoDB_Cluster.sh -nt ./AmazonWebServices_ArangoDB_Cluster.sh ] || [ Docker/ArangoDBClusterWithDocker.sh -nt ./AmazonWebServices_ArangoDB_Cluster.sh ] ; then
+  echo 'You almost certainly have forgotten to say "make" to assemble this'
+  echo 'script from its parts in subdirectories. Stopping.'
+  exit 1
+fi
+
 trap "kill 0" SIGINT
 
 #CoreOS AWS Image List
